Code cleanup

+++ src/memory.c
@@ -2,11 +2,11 @@
 #include <uv.h>
 
 uv_buf_t memory_uv_alloc(uv_handle_t* handle, size_t size) {
-	return uv_buf_init(malloc(size), size);
+	return uv_buf_init(malloc(size), (unsigned int)size);
 }
 
 uv_buf_t memory_alloc(size_t size) {
-	return uv_buf_init(malloc(size), size);
+	return uv_buf_init(malloc(size), (unsigned int)size);
 }
 
 uv_buf_t memory_realloc(uv_buf_t buf, size_t new_size) {
